requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
supabase
python-dotenv
